+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 - Modern JavaScript From The Beginning (Udemy)
   - Section 3: DOM Manipulation & Events
     - Removing & Replacing Elements
+    - Event Listeners & The Event Object
 - An example with removing and replacing of elements was added.
+- An example with event listeners was added.
 
 ## August 18, 2020
 - Modern JavaScript From The Beginning (Udemy)